refactor(oui-criteria): refactor oui-criteria-container to oui-criteria
