+++ packages/oui-angular/package.json
@@ -19,8 +19,7 @@
     "@ovh-ui/oui-chips": "^1.0.0",
     "@ovh-ui/oui-clipboard": "^1.0.0",
     "@ovh-ui/oui-collapsible": "^1.0.0",
-    "@ovh-ui/oui-criteria-adder": "^1.0.0",
-    "@ovh-ui/oui-criteria-container": "^1.0.0",
+    "@ovh-ui/oui-criteria": "^1.0.0",
     "@ovh-ui/oui-datagrid": "^1.0.0",
     "@ovh-ui/oui-dropdown": "^1.0.0",
     "@ovh-ui/oui-field": "^1.0.0",
