--- packages/oui-criteria-container/tests/index.js
@@ -1,7 +1,0 @@
-import "@ovh-ui/common/test-utils";
-
-loadTests(require.context("../src/", true, /.*((\.spec)|(index))$/));
-
-function loadTests (context) {
-    context.keys().forEach(context);
-}
