+++ packages/oui-criteria/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@ovh-ui/oui-criteria",
+  "version": "1.0.0",
+  "main": "./src/index.js",
+  "license": "BSD-3-Clause",
+  "author": "OVH SAS",
+  "dependencies": {
+    "@ovh-ui/oui-chips": "^1.0.0",
+    "@ovh-ui/oui-search": "^1.0.0"
+  }
+}
